==> S02 - Data Wrangling/HCKT02 - Data Wrangling/instructor_solution/1.local_files.ipynb <==
!ls -alh ../data

!head ../data/sample_submission.csv

Now we need to decompress the zip files.

! mkdir data

from zipfile import ZipFile

for i in range(1,4):
    # Create a ZipFile Object and load data_train_*.zip in it
    with ZipFile(f'../data/data_train_{i}.zip', 'r') as zipObj:
        # Extract all the contents of zip file in different directory
        zipObj.extractall(f'data/data_train_{i}')


### Data train 1

ls data/data_train_1/

import pandas as pd

df1 = pd.read_csv('data/data_train_1/datasource1.csv')

First error! It seems to be a unicodedecode error, this means the file is not unicode (utf8) valid. There are a huge number of encodings available, we can use [chardet](https://chardet.readthedocs.io/en/latest/usage.html) to detect the valid encoding.

To use it we just have to read the file in byte format (the `rb` flag). We only read the first 1000 bytes (so it's faster) and use `chardet.detect`.

import chardet

with open('data/data_train_1/datasource1.csv', 'rb') as fname:
    raw_data = fname.read()[:1000]
    
chardet.detect(raw_data)

df1 = pd.read_csv('data/data_train_1/datasource1.csv', encoding='windows-1251')

df1.head()

Let's set the id as the index for all the files.

df1 = df1.set_index('id')

Done! Let's go to the second file of this folder.

ls data/data_train_1/

It seems to be a json file...

with open('data/data_train_1/datasource2.json') as fname:
    df2 = pd.read_json(fname, lines=True)

df2.head()

Ok that didn't work, maybe a different orientation? Let's check the [documentation](https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.read_json.html) and preview the json.

# this shows the first 500 bytes of the file
! head -c 500 data/data_train_1/datasource2.json

with open('data/data_train_1/datasource2.json') as fname:
    df2 = pd.read_json(fname, orient='split')

df2.head()

Ok, let's go to the third file!

!head data/data_train_1/datasource3.csv

Simple, it's a csv but with a pipe `|` as a separator.

This is why doing a simple `!head` before running into pandas is super helpful!

df3 = pd.read_csv('data/data_train_1/datasource3.csv', sep='|')

df3.head()

df3 = df3.set_index('id')

Done! Let's move to the second folder.

#### data train 2

!ls -alh data/data_train_2/

Let's proceed with datasource4, it seems like an html file?

!head -n 20 data/data_train_2/datasource4.html

Let's try `pandas.read_html`!

# This takes a while to run...
df4 = pd.read_html('./data/data_train_2/datasource4.html', attrs = {'class': 'dataframe'})

`read_html` returns a list.

len(df4)

There is only one table on the html, so we get the first element of the list.

df4 = df4[0]

df4.head()

df4.columns

It seems it is a  multiindex!, we need to get the id as an index.

df4.columns = df4.columns.droplevel(level=0)

df4.head()

df4.columns = ['id', 'isusingmultipledevices', 'freereturn', 'userid']

df4.head()

df4 = df4.set_index('id')

Done! Let's move to file 5.

# this might crash your browser
# !head data/data_train_2/datasource5.csv

Ok, this file disguises itself as a csv but its actually a json! A good reminder that file extensions don't mean anything!

df5 = pd.read_json('data/data_train_2/datasource5.csv')

df5.head()

Done! Let's go to file 6.

!ls -alh data/data_train_2/

Well, it seems that there is no file 6 here, let's go to file 7 then.

!head data/data_train_2/datasource7.csv

Ok, we see this seems to be a csv that someone has messed with. Since csvs are basically text files this happens sometimes

No problem, we will skip the header and we will remove the columns that are useless and have to be deleted (the file is asking very nicely after all...).

df7 = pd.read_csv('data/data_train_2/datasource7.csv', 
                  header=None, 
                  skiprows=1,
                  usecols=[0,1,2,3],
                  names=['id','countrycode','countryoforigin','userfraudstatus']
                 )

df7.head()

df7 = df7.set_index('id')

Done!

### Folder 3

!ls -alh data/data_train_3/

Ok, here is the datasource6, which seems to be another folder?

# try running it without the "| head -10" to see all the files
!ls -alh data/data_train_3/datasource6/ | head -10

Ok, seems to be a lot of csv files. Since they have similar names, they are probably created by some automatic process. This is good because it means that we can use the same function to read all of the files.

!head data/data_train_3/datasource6/1000.csv

We will create a list of dataframes and will concatenate them.

from pathlib import Path

datasource6 = Path('data/data_train_3/datasource6/')

# this takes a while to run
ind_dfs = []
for file in datasource6.glob('*'):
    ind_dfs.append(pd.read_csv(file))

df6 = pd.concat(ind_dfs)

df6.shape

df6.head()

df6 = df6.set_index('id')

Done!

Now we merged them all!

files_df = df1.join(df2).join(df3).join(df4).join(df5).join(df6).join(df7)

files_df.head()

And output the resulting dataframe to a pickle in the `data/clean` directory!

! mkdir data/clean

files_df.to_pickle('data/clean/files_df.pkl')
