==> S02 - Data Wrangling/HCKT02 - Data Wrangling/instructor_solution/2.scraping.ipynb <==
import pandas as pd

We start scraping the website.

url = 'https://scrap-me.herokuapp.com/items?page={page}'

# this takes a while to run
scraped_dfs = []
page = 1
while 1:
    print(page)
    scraped_dfs.append(pd.read_html(url.format(page=page))[0])
    page +=1

Ok so there seems to be 400 pages

len(scraped_dfs)

scraped_df = pd.concat(scraped_dfs)

scraped_df.head()

scraped_df = scraped_df.set_index('id')

scraped_df.head()

scraped_df.to_pickle('data/clean/scraped.pkl')
