==> S02 - Data Wrangling/HCKT02 - Data Wrangling/instructor_solution/3.api.ipynb <==
Here we will retrieve the api information.

import requests
import pandas as pd
from tqdm import tqdm_notebook

base_url = 'https://y29rdnycjd.execute-api.eu-west-1.amazonaws.com/dev/'

r = requests.get(base_url)

r.ok

We see we can connect, so we don't need any authentication.

By checking the [api swagger doc](https://y29rdnycjd.execute-api.eu-west-1.amazonaws.com/dev/swagger.json), we see there is an endpoint named `missingdata` to which we pass an id and returns the missing columns.

endpoint_url = base_url + 'missingdata/{id}'
endpoint_url

sample_id = 'a0ef54d140f90a214b247122aa904a34'

r = requests.get(endpoint_url.format(id=sample_id))

r.request.url

r.content

So the endpoint works, but that id is not found.

Now we will read one of the dataframes we already created.

files_df = pd.read_pickle('data/clean/files_df.pkl')

files_df.columns

So I asked the instructors how to know which ids are in the API, I did it so I didn't have to blast the API with unnecessary requests.

It turns out, the rows that are in the API are those where column values are `API`.

ids_api = files_df[files_df.tierafterorder == 'API'].index.values

ids_api.shape

def get_missing_data(row_id):
    r = requests.get(endpoint_url.format(id=row_id))
    if r.ok:
        return r.json()
    else:
        return None

get_missing_data(ids_api[0])

It works!

Now we just need to do all of the requests. We will use [dask](dask.org) to speed up the process (see how to install it [here](https://docs.dask.org/en/latest/install.html)).

import dask
from dask.distributed import Client
client = Client()
client

data = [dask.delayed(get_missing_data)(row_id) for row_id in  ids_api]

res = dask.compute(data)

len(res[0])

res[0][0]

If you don't want to use dask, you can always do it in plain python! To do it, uncomment the code below.

#all_data = []
#for i in tqdm_notebook(ids_api):
#    all_data.append(get_missing_data(i))

Make a dataframe out of the json responses.

api_df = pd.DataFrame(res[0])

api_df.head()

api_df['id'] = ids_api

api_df = api_df.set_index('id')

api_df.to_pickle('data/clean/api_df.pickle')
