==> S02 - Data Wrangling/HCKT02 - Data Wrangling/instructor_solution/4.database.ipynb <==
On the datasources document for the hackathon, it says that the database is the one we used before, but using a new schema.

host = 'data-wrangling-batch3.cl9uj9cucww7.eu-west-1.rds.amazonaws.com'
port = 5432
database = 'datawrangling'
user = 'ldsa_student'
password = 'JGIYc6jD' # replace with the datab ase password!

We will use sqlalchemy to create a database connection (psycopg2 would work as well).

from sqlalchemy import create_engine

db_string = f"postgres://{user}:{password}@{host}:{port}/{database}"

db = create_engine(db_string)

conn = db.connect()

Check the tables under the schema we care about.

conn.execute("""
SELECT
   *
FROM
   pg_catalog.pg_tables
WHERE
   schemaname != 'pg_catalog'
AND schemaname != 'information_schema';
""").fetchall()

So we see that under the schema `hackathon_students` there are 2 tables: `targets` and `storeid`.

import pandas as pd

targets = pd.read_sql('select * from hackathon_students.targets', conn)

targets.head()

storeids = pd.read_sql('select * from hackathon_students.storeid', conn)

storeids.head()

targets = targets.set_index('id')
storeids = storeids.set_index('id')

targets.to_pickle('data/clean/targets.pkl')
storeids.to_pickle('data/clean/storeids.pkl')
