==> S02 - Data Wrangling/HCKT02 - Data Wrangling/instructor_solution/5.putting_all_together.ipynb <==
Now we have to put together the different datasets.

import pandas as pd

files_df = pd.read_pickle("data/clean/files_df.pkl")

We remove from the dataframe those rows whose origin is the website (these rows have `WEBSITE` in all the values) and those that come from the API (these rows have `API` in all the values).

files_df.shape

website_ids = files_df[files_df.tierafterorder.isin(['WEBSITE'])].index
api_ids = files_df[files_df.tierafterorder.isin(['API'])].index

files_df = files_df[-files_df.tierafterorder.isin(['WEBSITE', 'API'])]

files_df.shape

files_df.tierafterorder.value_counts()

scraped_df = pd.read_pickle('data/clean/scraped.pkl')

scraped_df.shape

scraped_df = scraped_df.loc[website_ids]

scraped_df.shape

scraped_df.head()

targets_df = pd.read_pickle('data/clean/targets.pkl')
storeid_df = pd.read_pickle('data/clean/storeids.pkl')

api_df = pd.read_pickle('data/clean/api_df.pickle')

We concat the dataframes that have different ids.

train_df = pd.concat(
    [
        pd.concat([files_df, api_df], sort=True),
        scraped_df
    ],
    sort=True
)

train_df.shape

Now we join the files that share an index.

train_df = train_df.drop(columns=['returned', 'storeid']).join(targets_df).join(storeid_df)

train_df.shape

train_df.to_pickle('data/clean/train_df_merged.pkl')
