==> S02 - Data Wrangling/HCKT02 - Data Wrangling/instructor_solution/6.test_dataset.ipynb <==
Now we will load the test file. This is important because when we deal with messy data we need to make sure that the columns in the train file exist at prediction time!

import pandas as pd

ls ../data/

test_df = pd.read_csv('../data/data_test.zip')

Ok, it seems that the test data is also split in Excel files. This means we need to decompress them and read one by one into `data/data_test` dir.

! mkdir data/data_tes

from zipfile import ZipFile

with ZipFile(f'../data/data_test.zip', 'r') as zipObj:
    # Extract all the contents of zip file in different directory
    zipObj.extractall(f'data/data_test')

!ls data/data_test

!ls data/data_test/transformed_data/

**Enough!!!!**

# You might not have this command if you're in a windows or mac
!tree data/data_test

!ls data/data_test/transformed_data/data_test/test

from pathlib import Path

test = Path('data/data_test/transformed_data/data_test/test/')

Now we iterate through every excel file. We can even use [tqdm](https://github.com/tqdm/tqdm) to get a progress indicator!

from tqdm import tqdm_notebook

files = list(test.glob('*'))
ind_dfs = []

for file in tqdm_notebook(files, total=len(files)):
    ind_dfs.append(pd.read_excel(file))

We append the same way we did with the datasource6.

test_df = pd.concat(ind_dfs)

test_df.head()

And save to pickle 

test_df.to_pickle('data/clean/test_df.pickle')
