==> S02 - Data Wrangling/HCKT02 - Data Wrangling/instructor_solution/7.baseline_model.ipynb <==
Here we will train a baseline model.
We will first load the test file, this is important because when we deal with messy data we need to make sure that the columns in the train file exist at prediction time!

import pandas as pd
import numpy as np

test_df = pd.read_pickle("data/clean/test_df.pickle")

test_df.columns

corr = test_df.corr()
corr[corr>0.99]

*Note*, you can also use [pandas-profiling!](https://github.com/pandas-profiling/pandas-profiling)

#from pandas_profiling import ProfileReport
#report = ProfileReport(test_df)
#report

We will remove  the following columns because they are highly correlated with other columns:
- countrycode (0.999918 corr with country)
- style (0.99971 corr with designer)

test_df.head()

`test_df.orderdate_gmt` is a datetime, we can parse it in later iterations but for now we remove it.

test_df = test_df.drop(columns=['countrycode', 'orderdate_gmt', 'style'])

Now we load the training data that we previously cleaned.

train_df = pd.read_pickle('data/clean/train_df_merged.pkl')

#train_report = ProfileReport(train_df)
#train_report

We evaluate cardinality of categoricals.

cardinality = train_df.describe(exclude=np.number).T

cardinality

We see the column `ddprate` is a numerical (top value is 0.0) but it's not considered a category.

train_df.ddprate.value_counts(True)

So it seems like it's a number, but somehow got defined as a categorical. Also we see there are two different kinds of 0.0?

train_df.ddprate.value_counts().head().index.values

train_df.ddprate = pd.to_numeric(train_df.ddprate, errors='coerce')

cardinality[cardinality.unique>100000]

We see there are a bunch of columns that are almost ids, we should use a hashingencoder to deal with them but for now we remove them.

train_df = train_df.drop(columns=['orderportalid', 'userid', 'style', 'productid',
                                  'orderdate_gmt', 'designer'])

target_name = 'returned'
y = train_df[target_name]
X = train_df.drop(columns=[col for col in train_df if col not in test_df.columns])

For the moment we will remove `orderportalid` since it has high cardinality.

X.describe(exclude=np.number).T

We save it for later.

X.assign(returned=y).reset_index().to_csv('train.csv', index=False)

# Baseline Model

from sklearn.linear_model import LogisticRegression
import numpy as np
from sklearn.pipeline import make_pipeline
from category_encoders import OneHotEncoder
from sklearn.model_selection import train_test_split
from sklearn.impute import SimpleImputer

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.1)

print(X_train.shape)

pipeline = make_pipeline(
 OneHotEncoder(verbose=1),
    SimpleImputer(),
    LogisticRegression(verbose=1)
)

For the first run let's just train on a sample, to make it faster.

pipeline.fit(X_train.head(100000), y_train.values[:100000])

preds = pipeline.predict(X_test)

from sklearn.metrics import roc_auc_score

roc_auc_score(preds, y_test)

X_test.shape

We load the test data.

test_df = pd.read_pickle('data/clean/test_df.pickle')
test_ids = test_df.id.values

test_df = test_df[[col for col in test_df.columns  if col in X.columns ]]

test_df.columns

test_preds = pipeline.predict_proba(test_df)[:,1]

baseline = pd.DataFrame({
    "id":test_ids,
    "returned":  test_preds
    
})

baseline.to_csv('baseline.csv', index=False)

We got ROCAUC of 0.68 on the leaderboard with 100k rows. Let's try with the whole dataset.

# This takes a while to run
pipeline.fit(X_train, y_train)

preds = pipeline.predict(X_test)

roc_auc_score(preds, y_test)

test_preds2 = pipeline.predict_proba(test_df)[:,1]

baseline2 = pd.DataFrame({
    "id":test_ids,
    "returned":  test_preds2
    
})
baseline2.to_csv('baseline.2.csv', index=False)

print('done')

This got a 0.773!

We save the test df just in case

test_df['id'] = test_ids
test_df.to_csv('test.csv', index=False)
